Refactoring code to account for new utils dir and schema file
