+++ suzieq/engines/pandas/bgp.py
@@ -2,7 +2,7 @@ import pandas as pd
 import numpy as np
 
 from .engineobj import SqPandasEngine
-from suzieq.utils import build_query_str, humanize_timestamp
+from suzieq.shared.utils import build_query_str, humanize_timestamp
 
 
 class BgpObj(SqPandasEngine):
