+++ suzieq/engines/pandas/path.py
@@ -10,7 +10,7 @@ import pandas as pd
 
 from suzieq.exceptions import EmptyDataframeError
 from .engineobj import SqPandasEngine
-from suzieq.utils import expand_nxos_ifname, MAX_MTU
+from suzieq.shared.utils import expand_nxos_ifname, MAX_MTU
 
 # TODO: What timestamp to use (arpND, mac, interface, route..)
