+++ suzieq/poller/services/arpnd.py
@@ -1,7 +1,8 @@
 from suzieq.poller.services.service import Service
 import re
 import numpy as np
-from suzieq.utils import (convert_macaddr_format_to_colon, expand_ios_ifname)
+from suzieq.shared.utils import (
+    convert_macaddr_format_to_colon, expand_ios_ifname)
 
 
 class ArpndService(Service):
