+++ suzieq/poller/services/device.py
@@ -1,7 +1,7 @@
 from datetime import timedelta, datetime
 
 from suzieq.poller.services.service import Service
-from suzieq.utils import get_timestamp_from_junos_time
+from suzieq.shared.utils import get_timestamp_from_junos_time
 from dateparser import parse
