+++ suzieq/poller/services/routes.py
@@ -3,8 +3,8 @@ from dateparser import parse
 from datetime import datetime
 
 from suzieq.poller.services.service import Service
-from suzieq.utils import (expand_nxos_ifname, get_timestamp_from_cisco_time,
-                          get_timestamp_from_junos_time)
+from suzieq.shared.utils import (expand_nxos_ifname, get_timestamp_from_cisco_time,
+                                 get_timestamp_from_junos_time)
 
 
 import numpy as np
